+++ packages/vault/src/type.ts
@@ -1,4 +1,5 @@
 export interface Store {
-  setItem(key: string, value: string): void;
-  getItem(key: string): string;
+  set(key: string, value: string): void;
+  get(key: string): string;
+  keys(): string[];
 }