+++ packages/api/CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - [#179](https://github.com/mesg-foundation/js-sdk/pull/179) Add LCD API for all reading (execution, instance, ownership, process, runner, service)
 - [#179](https://github.com/mesg-foundation/js-sdk/pull/179) Migrate resolver to use LCD server
+- [#183](https://github.com/mesg-foundation/js-sdk/pull/183) Add account import from a mnemonic
 
 #### Bug fixes