+++ packages/vault/CHANGELOG.md
@@ -4,4 +4,7 @@
 
 #### Breaking Changes
 #### Improvements
+
+- [#183](https://github.com/mesg-foundation/js-sdk/pull/183) Add localstorage store and add `keys` function to the store
+
 #### Bug fixes